>Fixed a bug to do with parsing JAMS data that contains user logs. >Fixed a floating point precision error.

+++ annotation-tool/annotation-tool/classes/parsers/FileParser.cs
@@ -842,7 +842,7 @@ namespace AnnotationTool
                     int occurrenceIndex = Int32.Parse(data.value["occurrence_id"]);
                     int confidence = Int32.Parse(data.confidence);
                     int start = Int32.Parse(data.time);
-                    int duration = Int32.Parse(data.duration);
+                    double duration = Double.Parse(data.duration);
 
                     while (patternIndex >= patterns.Count)
                     {
